refactor: struct configurations
