+++ Server/src/main/resources/config.properties
@@ -1,4 +1,4 @@
-# configuration.properties
+# configurations.properties
 redis.host=localhost
 refresh.token.key.prefix=YLaLauRoiKhongGapChacCung30NgayRoiHe
 access.token.key.prefix=YLaLauRoiKhongGapChacCung30NgayRoiHe
